move test case to DataFrameAggregateSuite
